+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Chapter inherits Narrative
 - Plot attribute proxy for chapters and subplots
 - Entity#leave
+- Persistent subplots
 
 ## 3.4.0 - September 10, 2024
 - Chapters